+++ frontend/src/components/MeetingRoom/ParticipantListItemMenu/ParticipantListItemMenu.tsx
@@ -1,14 +1,8 @@
 import React from 'react';
 import { IconButton, Menu } from '@mui/material';
-import PushPinIcon from '@mui/icons-material/PushPin';
 import MoreVertIcon from '@mui/icons-material/MoreVert';
-import MenuItem from '@mui/material/MenuItem';
-import ListItemText from '@mui/material/ListItemText';
-import ListItemIcon from '@mui/material/ListItemIcon';
-import PushPinOffIcon from '../../Icons/PushPinOffIcon';
 import { SubscriberWrapper } from '../../../types/session';
-import useSessionContext from '../../../hooks/useSessionContext';
-import truncateString from '../../../utils/truncateString';
+import ParticipantPinMenuItem from './ParticipantPinMenuItem';
 
 type ParticipantListItemMenuToggleButtonProps = {
   participantName: string;
@@ -19,33 +13,13 @@ const ParticipantListItemMenuToggleButton = ({
   participantName,
   subscriberWrapper,
 }: ParticipantListItemMenuToggleButtonProps) => {
-  const { isPinned, id } = subscriberWrapper;
-  const { isMaxPinned, pinSubscriber } = useSessionContext();
   const [anchorEl, setAnchorEl] = React.useState<null | HTMLElement>(null);
   const open = Boolean(anchorEl);
   const handleClick = (event: React.MouseEvent<HTMLButtonElement>) => {
     setAnchorEl(event.currentTarget);
   };
-  const isDisabled = !isPinned && isMaxPinned;
   const handleClose = () => {
     setAnchorEl(null);
-  };
-
-  const trimmedName = truncateString(participantName, 22);
-
-  const getText = () => {
-    if (isPinned) {
-      return `Unpin ${trimmedName}`;
-    }
-    if (isMaxPinned) {
-      return `You can't pin any more tiles`;
-    }
-    return `Pin ${trimmedName}`;
-  };
-  const handlePinClick = () => {
-    if (!isDisabled) {
-      pinSubscriber(id);
-    }
   };
 
   return (
@@ -62,16 +36,10 @@ const ParticipantListItemMenuToggleButton = ({
           horizontal: 250,
         }}
       >
-        <MenuItem disabled={isDisabled} sx={{ width: '280px' }} onClick={handlePinClick}>
-          <ListItemIcon>
-            {isPinned ? (
-              <PushPinOffIcon fontSize="small" sx={{ color: 'black' }} />
-            ) : (
-              <PushPinIcon fontSize="small" sx={{ color: 'black' }} />
-            )}
-          </ListItemIcon>
-          <ListItemText sx={{ maxWidth: '200px' }}>{getText()}</ListItemText>
-        </MenuItem>
+        <ParticipantPinMenuItem
+          subscriberWrapper={subscriberWrapper}
+          participantName={participantName}
+        />
       </Menu>
     </>
   );
